requirements.txt
# Core AI and ML Libraries
openai>=1.12.0
anthropic>=0.18.1
google-generativeai>=0.3.2
langchain>=0.1.10
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.6

# Vector Databases
chromadb>=0.4.22

# UI Framework
streamlit>=1.31.1
streamlit-option-menu>=0.3.12

# Database
supabase>=2.3.0
psycopg2-binary>=2.9.9

# PDF Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
pdfplumber>=0.10.3

# API Integrations
google-api-python-client>=2.110.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
notion-client>=2.2.1
gspread>=5.12.4

# Utilities
python-dotenv>=1.0.1
pandas>=2.2.1
numpy>=1.26.3
tiktoken>=0.5.2
sentence-transformers>=2.3.1
beautifulsoup4>=4.12.3
requests>=2.31.0

